requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==4.0.4
django-heroku==0.3.1
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-cloud-dataproc==4.0.2
googleapis-common-protos==1.56.1
grpcio==1.46.1
grpcio-status==1.46.1
gunicorn==20.1.0
heroku-cli-wrapper==0.0.8
httplib2==0.20.4
idna==3.3
javascript==1!1.0.0
proto-plus==1.20.3
protobuf==3.20.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
requests==2.27.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.1.0
